Fix thread exhaustion bug
The fusionCallback would block waiting to acquire the lock. This caused
a thread exhaustion and crash in cases where messages were not being
processed fast enough.

To fix this issue, the Depth/RGB/Pose callbacks just add the incoming
data into circular buffers, without calling fusionCallback. The
runPipelineOnce (fusionCallback) is then called repeatedly, ensuring it
is only run by a single thread at a time. The crucial part is that
instead of blocking if another thread holds the lock, it just returns
and will be called later.